Stage 4.8.4d: widen kplib's pointer carriers + cascade
kplib.c carried framebuffer addresses through 4 layers of int32_t
storage:
  1. public APIs:    int32_t kprender(_, int32_t frameptr, ...)
                      void    kpzload(_, int32_t *pic, ...)
  2. internal calls: each kp*rend(_, int32_t daframeplace, ...)
  3. file-scope:     static int32_t frameplace; static int32_t nfplace;
  4. function-local: int32_t p, pp, yoff, i, j, ...

On x86_64 every layer silently truncated. Widened all four to
intptr_t so the host pointer survives end-to-end.

API-visible:
  - kprender    frameptr  int32_t   → intptr_t
  - kpzload    *pic       int32_t * → intptr_t *

Internal renderer parameters (8 functions: kpngrend, kpegrend,
kgifrend, kcelrend, ktgarend, kbmprend, kpcxrend, kddsrend):
  - daframeplace / frameptr  → intptr_t

Locals:
  - putbuf p, rgbhlineasm/pal8hlineasm p, kgifrend yoff,
    ktgarend p, kpcxrend i+p, kddsrend (renamed `j` → `fbrow`
    where it carried an address; kept `j` int32_t for its
    DXT loop-counter use), yrbrend p+pp.

Forced cascade in voxlap5.c:
  - skypic, nskypic, skyoff (sky texture base addresses written by
    kpzload) → intptr_t.

kpgetdim pointer arithmetic at lines 2488/2498 was using
`(uint32_t)ptr - (uint32_t)ptr` to compute byte offsets — same
truncation. Replaced with `(uintptr_t)((const char *)ptr - buf)`
so the comparison stays well-defined on LP64.

Bit-identical on Windows MSVC x86 (intptr_t == int32_t there).
Linux GCC/Clang x86_64: int-to-pointer-cast warnings 69 → 36;
kplib.c is now fully clean. Oracle still SIGSEGVs — every
remaining truncation lives in voxlap5.c kv6/sprite/malloc paths
(4.8.4e).

+++ voxlap/voxlap5.h
@@ -290,10 +290,10 @@ VOXLAP_API extern void voxrestore ();
 VOXLAP_API extern void voxredraw ();
 
 	//High-level (easy) picture loading function:
-VOXLAP_API extern void kpzload (const char *, int32_t *, int32_t *, int32_t *, int32_t *);
+VOXLAP_API extern void kpzload (const char *, intptr_t *, int32_t *, int32_t *, int32_t *);
 	//Low-level PNG/JPG functions:
 VOXLAP_API extern void kpgetdim (const char *, int32_t, int32_t *, int32_t *);
-VOXLAP_API extern int32_t kprender (const char *, int32_t, int32_t, int32_t, int32_t, int32_t, int32_t, int32_t);
+VOXLAP_API extern int32_t kprender (const char *, int32_t, intptr_t, int32_t, int32_t, int32_t, int32_t, int32_t);
 
 	//ZIP functions:
 VOXLAP_API extern int32_t kzaddstack (const char *);
